+++ app/controllers/mapping_controller.rb
@@ -150,6 +150,13 @@ class MappingController < ApplicationController
         target_system_id = '6b31a67a-7e6d-57c0-8609-52912076fce8'
         target_version_id = 'b5165f68-b934-5c79-ac71-bd5375f7c809'
 
+        # get the options to populate the Equivalence Type dropdown
+        equivalence_options = [{value: '', label: 'No Restrictions'}]
+
+        get_direct_children($isaac_metadata_auxiliary['MAPPING_QUALIFIERS']['uuids'].first[:uuid], true, true).each do |option|
+            equivalence_options << {value: option[:concept_id], label: option[:text]}
+        end
+
         @set_id = params[:set_id]
 
         if @set_id &&  @set_id != ''
@@ -232,7 +239,7 @@ class MappingController < ApplicationController
                 # setup the intrinsic map item fields
                 source_info = {id: 'DESCRIPTION_SOURCE', description: 'Mapping Source Concept', order: 0, data_type: 'UUID', required: true, text: 'Mapping Source Concept', removable: false, display: true, component_type: 'SOURCE'}
                 target_info = {id: 'DESCRIPTION_TARGET', description: 'Mapping Target Concept', order: 1, data_type: 'UUID', required: true, text: 'Mapping Target Concept', removable: false, display: true, component_type: 'TARGET'}
-                qualifier_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: 2, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE'}
+                qualifier_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: 2, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE', options: equivalence_options}
 
                 # load the intrinsic map item fields into our return mapset variable
                 @map_set['item_field_DESCRIPTION_SOURCE'] = source_info
@@ -297,7 +304,7 @@ class MappingController < ApplicationController
                     elsif field.id == 'DESCRIPTION' && field.componentType.enumName == 'EQUIVALENCE_TYPE'
 
                         @map_set[:item_fields] << 'DESCRIPTION_EQUIVALENCE_TYPE'
-                        field_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: computed_field_index, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE'}
+                        field_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: computed_field_index, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE', options: equivalence_options}
                         @map_set['item_field_DESCRIPTION_EQUIVALENCE_TYPE'] = field_info
                         computed_field_index += 1
 
@@ -378,7 +385,7 @@ class MappingController < ApplicationController
             # setup the intrinsic map item fields
             source_info = {id: 'DESCRIPTION_SOURCE', description: 'Mapping Source Concept', order: nil, data_type: 'UUID', required: true, text: 'Mapping Source Concept', removable: false, display: true, component_type: 'SOURCE'}
             target_info = {id: 'DESCRIPTION_TARGET', description: 'Mapping Target Concept', order: nil, data_type: 'UUID', required: true, text: 'Mapping Target Concept', removable: false, display: true, component_type: 'TARGET'}
-            qualifier_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: nil, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE'}
+            qualifier_info = {id: 'DESCRIPTION_EQUIVALENCE_TYPE', description: 'Equivalence Type', order: nil, data_type: 'SELECT', required: true, text: 'Equivalence Type', removable: false, display: true, component_type: 'EQUIVALENCE_TYPE', options: equivalence_options}
 
             # load the intrinsic map item fields into our return mapset variable
             @map_set['item_field_DESCRIPTION_SOURCE'] = source_info
@@ -427,6 +434,8 @@ class MappingController < ApplicationController
         #show_inactive = params[:show_inactive]
         results = {column_definitions: column_definitions}
         item_data = []
+        source_name = ''
+        target_name = ''
 
         items = MappingApis::get_mapping_api(action: MappingApiActions::ACTION_ITEMS, uuid_or_id: set_id,  additional_req_params: {coordToken: coordinates_token, allowedStates: view_params['statesToView'], expand: 'referencedDetails,comments '}) # CommonRest::CacheRequest => false
 
@@ -437,6 +446,8 @@ class MappingController < ApplicationController
         items.each do |item|
 
             item_hash = {}
+            source_name = ''
+            target_name = ''
 
             item_hash[:item_id] = item.identifiers.uuids.first
             item_hash[:state] = item.mappingItemStamp.state.enumName
@@ -478,7 +489,15 @@ class MappingController < ApplicationController
 
                     # check to see if computedDisplayFields is present, otherwise just use the UUID for the value
                     if item.computedDisplayFields
+
                         item_hash[column_definition[:id] + '_display'] = item.computedDisplayFields[column_definition[:order].to_i].value
+
+                        # if this is the source or target concept record the display name in the return object for convenient access since these columns aren't always in a predefined order
+                        if column_definition[:id] == 'DESCRIPTION_SOURCE'
+                            source_name = item.computedDisplayFields[column_definition[:order].to_i].value
+                        elsif column_definition[:id] == 'DESCRIPTION_TARGET' && item.computedDisplayFields[column_definition[:order].to_i].value != nil
+                            target_name = item.computedDisplayFields[column_definition[:order].to_i].value
+                        end
                     else
                         item_hash[column_definition[:id] + '_display'] = item_hash[column_definition[:id]]
                     end
@@ -517,6 +536,9 @@ class MappingController < ApplicationController
                         item_hash[column_definition[:id]] = html_escape(field.data)
                     end
                 end
+
+                # record the item display name in the item hash for convenient access since these columns aren't always in a predefined order
+                item_hash[:item_name] = source_name + ' - ' + target_name
             end
 
             item_data << item_hash